@@ -39,7 +39,7 @@ dependencies:
   # Agregadas dependencias necesarias para tu proyecto
   provider: ^6.0.0
   flutter_date_pickers: ^0.4.3
-  supabase_flutter: ^2.5.2
+  supabase_flutter: ^2.9.1
   pdf: ^3.10.8
   printing: ^5.12.0
   share_plus: ^11.0.0
